# Client/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "v1",
		"title": "Meyer USPS API"
	},
	"paths": {
		"/api/usps/verify": {
			"get": {
				"tags": [ "Address" ],
				"operationId": "VerifyAddress",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "firmname",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "address1",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "address2",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "city",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "state",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip5",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip4",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "urbanization",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/Address" }
					},
					"400": {
						"description": "Bad Request",
						"schema": { "type": "string" }
					}
				}
			}
		},
		"/api/usps/citystatelookup": {
			"get": {
				"tags": [ "Address" ],
				"operationId": "LookupCityState",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "zip51",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip52",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip53",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip54",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "zip55",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"uniqueItems": false,
							"type": "array",
							"items": { "$ref": "#/definitions/CityState" }
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": { "type": "string" }
					}
				}
			}
		},
		"/api/usps/zipcodelookup": {
			"get": {
				"tags": [ "Address" ],
				"operationId": "LookupZipCode",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "address1",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "address2",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "city",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "firmname",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "state",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "urbanization",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/ZipCode" }
					}
				}
			}
		},
		"/api/usps/Tracking": {
			"get": {
				"tags": [ "Tracking" ],
				"operationId": "Tracking",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "trackId1",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "trackId2",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "trackId3",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "trackId4",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "trackId5",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/Track" }
					},
					"400": {
						"description": "Bad Request",
						"schema": { "type": "string" }
					}
				}
			}
		},
		"/api/usps/TrackByEmail": {
			"get": {
				"tags": [ "Tracking" ],
				"operationId": "TrackByEmail",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/CityState" }
					},
					"400": {
						"description": "Bad Request",
						"schema": { "type": "string" }
					}
				}
			}
		},
		"/api/usps/ProofOfDelivery": {
			"get": {
				"tags": [ "Tracking" ],
				"operationId": "ProofOfDelivery",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "address1",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "address2",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "city",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "firmname",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "state",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "urbanization",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/ZipCode" }
					}
				}
			}
		},
		"/api/usps/ReturnReceiptElectronic": {
			"get": {
				"tags": [ "Tracking" ],
				"operationId": "ReturnReceiptElectronic",
				"consumes": [],
				"produces": [ "application/json" ],
				"parameters": [
					{
						"name": "address1",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "address2",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "city",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "firmname",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "state",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "urbanization",
						"in": "query",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": { "$ref": "#/definitions/ZipCode" }
					}
				}
			}
		}
	},
	"definitions": {
		"Address": {
			"type": "object",
			"properties": {
				"firmName": { "type": "string" },
				"address1": { "type": "string" },
				"address2": { "type": "string" },
				"city": { "type": "string" },
				"cityAbbreviation": { "type": "string" },
				"state": { "type": "string" },
				"zip5": { "type": "string" },
				"zip4": { "type": "string" },
				"deliveryPoint": { "type": "string" },
				"carrierRoute": { "type": "string" },
				"footnotes": { "type": "boolean" },
				"dpvConfirmation": { "type": "boolean" },
				"dpvcmra": { "type": "boolean" },
				"dpvFootnotes": { "type": "string" },
				"business": { "type": "boolean" },
				"centralDeliveryPoint": { "type": "boolean" },
				"vacant": { "type": "boolean" },
				"urbanization": {
					"type": "string",
					"readOnly": true
				},
				"id": { "type": "string" },
				"error": { "type": "string" }
			}
		},
		"CityState": {
			"type": "object",
			"properties": {
				"zip5": { "type": "string" },
				"city": { "type": "string" },
				"state": { "type": "string" },
				"id": { "type": "string" },
				"error": { "type": "string" }
			}
		},
		"ZipCode": {
			"type": "object",
			"properties": {
				"firmName": { "type": "string" },
				"address1": { "type": "string" },
				"address2": { "type": "string" },
				"city": { "type": "string" },
				"state": { "type": "string" },
				"zip5": { "type": "string" },
				"zip4": { "type": "string" },
				"id": { "type": "string" },
				"error": { "type": "string" }
			}
		},
		"Track": {
			"type": "object",
			"properties": {
				"trackSummary": { "type": "string" },
				"id": { "type": "string" },
				"error": { "type": "string" }
			}
		}
	},
	"tags": []
}